Consolidated unit tests and fixed install rules
