+++ reach_core/test/reach_study_utest.cpp
@@ -2,7 +2,9 @@
 #include <reach_core/utils.h>
 #include <reach_core/reach_study_comparison.h>
 
+#include <boost/algorithm/string.hpp>
 #include <boost/filesystem.hpp>
+#include <boost_plugin_loader/plugin_loader.hpp>
 #include <gtest/gtest.h>
 #include <random>
 #include <yaml-cpp/yaml.h>
@@ -36,6 +38,40 @@ reach::ReachDatabase createDatabase(const std::string& name = "reach_study")
   }
 
   return db;
+}
+
+template <typename PluginT>
+class PluginTest : public ::testing::Test
+{
+public:
+  PluginTest() : ::testing::Test()
+  {
+    loader.search_libraries_env = SEARCH_LIBRARIES_ENV;
+    boost::split(loader.search_libraries, PLUGIN_LIBRARIES, boost::is_any_of(":"), boost::token_compress_on);
+  }
+
+  boost_plugin_loader::PluginLoader loader;
+};
+
+// Test Implementations
+using Implementations = ::testing::Types<reach::EvaluatorFactory, reach::IKSolverFactory, reach::DisplayFactory,
+                                         reach::TargetPoseGeneratorFactory, reach::LoggerFactory>;
+
+TYPED_TEST_CASE(PluginTest, Implementations);
+
+TYPED_TEST(PluginTest, LoadPlugins)
+{
+  std::vector<std::string> declared_classes = this->loader.getAvailablePlugins(TypeParam::getSection());
+  EXPECT_GT(declared_classes.size(), 0);
+
+  // Loop over the available classes and attempt to load them
+  for (const std::string& cls : declared_classes)
+  {
+    std::cout << "Attempting to load '" << cls << "' class" << std::endl;
+    std::shared_ptr<TypeParam> plugin;
+    EXPECT_NO_THROW(plugin = this->loader.template createInstance<TypeParam>(cls));
+    EXPECT_NE(plugin, nullptr);
+  }
 }
 
 TEST(ReachStudy, ReachStudy)
